refactor: merge HintProcessor into VirtualMachine and remove circular dependencies
