+++ src/hints/hintName.ts
@@ -1,0 +1,5 @@
+/** Name to identify which hint is executed  */
+export enum HintName {
+  AllocSegment = 'AllocSegment',
+  TestLessThan = 'TestLessThan',
+}
